11 Сверстать страницу лидерборд, правки по ревью
