--- packages/client/src/pages/LeaderboardPage/types.tsx
@@ -1,10 +1,0 @@
-export type Leaderboard = {
-  place: number
-  playerName: string
-  playerAvatar: string
-  amount: number
-}
-
-export type LeaderboardListProps = {
-  list: Leaderboard[]
-}
