+++ packages/client/src/pages/MainPage/ui/MainPage.tsx
@@ -4,7 +4,7 @@ import { Route, Rule, Teammate } from '@pages/MainPage/types'
 import { Rules } from './Rules'
 import { Cta } from './Cta'
 import { Teammates } from '@pages/MainPage/ui/Teammates'
-import { Intro } from '@pages/MainPage/ui/Intro'
+import { Intro } from '@widgets/intro/Intro'
 import RulesImg1 from '@pages/MainPage/ui/assets/rules-img-1.svg'
 import RulesImg2 from '@pages/MainPage/ui/assets/rules-img-2.svg'
 import RulesImg3 from '@pages/MainPage/ui/assets/rules-img-3.svg'
