+++ packages/client/src/widgets/intro/Intro.tsx
